refactor: organizes the istio itests better
