+++ tests/integration/istio/__init__.py
@@ -1,0 +1,1 @@
+"""Istio integration tests package."""
