+++ tests/integration/istio/steps/__init__.py
@@ -1,0 +1,1 @@
+"""Step definitions for pytest-bdd tests."""
