+++ tests/integration/istio/tests/test_authorization_policies.py
@@ -3,4 +3,4 @@
 from pytest_bdd import scenarios
 
 # Load all scenarios from the authorization policies feature file
-scenarios("features/authorization-policies.feature")
+scenarios("../features/authorization-policies.feature")
