+++ tests/integration/istio/tests/test_istio_integration.py
@@ -3,4 +3,4 @@
 from pytest_bdd import scenarios
 
 # Load all scenarios from the feature file
-scenarios("features/istio-integration.feature")
+scenarios("../features/istio-integration.feature")
