+++ tests/integration/istio/tests/test_managed_mode.py
@@ -3,4 +3,4 @@
 from pytest_bdd import scenarios
 
 # Load all scenarios from the managed-mode feature file
-scenarios("features/managed-mode.feature")
+scenarios("../features/managed-mode.feature")
