+++ tox.ini
@@ -75,4 +75,4 @@ commands_pre =
 passenv =
     {[testenv]passenv}
 commands =
-    uv run --frozen pytest -v {toxinidir}/tests/integration/istio {posargs}
+    uv run --frozen pytest -v {toxinidir}/tests/integration/istio/tests {posargs}
